Adding auxiliary library
